+++ diagonalization/diagonalize.h
@@ -30,8 +30,8 @@ void printoutSVD(SVD svd);
 extern const int angles[14];
 
 void cordic(int* cos, int* sin, int angle);
-void dot_productM(int m1[M][M], int m2[M][M], int dest[M][M]);
-void transposeM(int source[M][M], int dest[M][M]);
-void print_matrix2(int matrix[2][2]);
-void print_matrixM(int matrix[M][M]);
-void print_descaled(int matrix[M][M]);
+void dot_productM(int16_t m1[M][M], int16_t m2[M][M], int16_t dest[M][M]);
+void transposeM(int16_t source[M][M], int16_t dest[M][M]);
+void print_matrix2(int16_t matrix[2][2]);
+void print_matrixM(int16_t matrix[M][M]);
+void print_descaled(int16_t matrix[M][M]);
